[registry] Disable token refresh thread when prefetch is enabled
Add a new RegistryConfig field called `enable_token_refresh` that will decide whether the registry refresh token thread is started on not.
It will be set to true by default but will be disabled when prefetch is enabled. The refresh thread is disabled in this case because it becomes useless once everything has been prefetched and it's unlikely that tokens would expire during the time it takes for the prefetch to run (as it would mean that regular image pulls  would be likely to fail as well because they download the layers in a similar time to the prefetch)

Signed-off-by: Baptiste Girard-Carrabin <baptiste.girardcarrabin@datadoghq.com>