[nydusify] Convert nydus images to/from local archives
This adds the ability to convert nydus images to and from local archive files using nydusify convert.
The main benefits are that images conversions can now be run in environments without access to a container registry (or internet for that matter), while still being able to use nydusify capabilities to build images with merge platforms or OCI referrer.
The feature is used with the flags --source-archive and --target-archive to specify the input and output archive files respectively. It's not necessary to specify both.
The --source and --target/--target-suffix flags remain necessary as they are needed to identify the images imported/exported in the archives.

Signed-off-by: Baptiste Girard-Carrabin <baptiste.girardcarrabin@datadoghq.com>