cherry picks for v2.3.8 dd.1 (#7)
* [BlobcacheMetrics] Increment entries_count when chunks become ready

The entries_count metric in BlobcacheMetrics was defined but never incremented, causing it to always report 0. This metric should track the number of ready chunks in the blob cache.

This change increments entries_count in two places:
1. In _update_chunk_pending_status() when set_ready_and_clear_pending() succeeds, indicating a chunk has been downloaded and persisted
2. During cache recovery when chunks are restored from disk on startup

Signed-off-by: Baptiste Girard-Carrabin <baptiste.girardcarrabin@datadoghq.com>

* [rafs] Allow configuration of prefetch batch_size with RAFS v6

The prefetch batch_size was hardcoded at 2MiB which prevented users from configuring it to control the amount of API calls made to the backend. This is now possible using the existing `batch_size` field in the prefetch configuration.

Signed-off-by: Baptiste Girard-Carrabin <baptiste.girardcarrabin@datadoghq.com>

---------

Signed-off-by: Baptiste Girard-Carrabin <baptiste.girardcarrabin@datadoghq.com>